Add missing pkgs for LVM service
While switching over Cinders BlockBox to build directly from
github.com/openstack/loci-cinder I noticed that there are a
couple of things needed still.

Since originally adding the loci-cinder lvm Dockerfile we've
switched cinder to use thin provisioning as a default.  As
a result, now we need to make sure the Dockerfile includes
adding thin tools for deb based images.

While looking at this I also noticed that the Centos image
is missing the tgt utils, so install the epel rep and the
scsi-target-utils to make centos work as well.

I realize we're workign towards a universal dockerfile and
will update things to work in the new model as well when
we're ready.

Change-Id: I655612b136fe7fe96e87ccf8b8e421d4f5fb3f4b